Fix for #7 - crash/unnecessary work when running BFG twice on same repo
The BFG attempts to translate any hex string it sees in commit messages,
so long as that hex string corresponds to an object id that the repo
actually contains.

If it has actually spotted a 'former commit' then this can easily blow
the stack, but even if it didn't, it would involve doing a lot of
unnecessary work, cleaning commits that are no longer referenced...
fixing stackexceptions is one thing, but this commit addresses the
unnecessary work by removing doing a git prune to remove all
unreferenced objects - so that legacy objct ids can not trigger work.

https://github.com/rtyley/bfg-repo-cleaner/issues/7

Former-commit-id: 42e1934f44756086da044671edfdb64c23edca36