Allow users to use regex, globs or literals if they specify them...
Supporting two sets of options for regexs and literals was just going
to get unwieldy, so now regex, globs or literals can be used pretty
much interchangeably, so long as the user uses an appropriate prefix
('regex:', 'glob:' or 'literal:').

Also tweaks to the command-line args, attempting to make them clearer.

--strict-object-checking is fairly redundant, so leaving it out until
scopt supports not listing it in 'usage' :-)