Use a faster charset detector, give ICU4J as a backup option
Faster charset detector is based on the JGIt awParseUtils.decode()
methods, which in turn are supposed to follow the behaviour of C-Git,
I think.

What's quite fun is that all the text-modifier tests pass with both
charset detectors (icu4j and new-quick-one) - because my tests only
involve replacing ASCII in the middle of other charsets at the moment.


Former-commit-id: 603e18829a1dddefc3606062183dd654d32bcfcc