Fix for #9 - clean commit messages even if they are on a clean branch
Scenario : A repo with two branches operating side-by-side. One branch
is completely clean (tree and parent-commit wise), but the other is
dirty. On the 'clean' branch, a human writes a commit message
referencing a commit from the dirty branch. This commit id must be
updated, even though it is contained in the text of an ostensibly
'clean' commit on a clean branch.


Former-commit-id: bda1d7a9a6703e0f72eddbcd33b5c6645e7d0067