Update to JGit (almost) 3.0...
The upgrade to JGit 3.0 moves a large chunk of functionality into the
org.eclipse.jgit.internal.* packages, but as far as The BFG is concerned
we can still get by using non-internal public interfaces.

The one exception to this is the BFG's packfile-scanning code (used to
find the largest blobs in a repo before starting the main run), which
uses file-storage related classes:

org.eclipse.jgit.internal.storage.file.{ObjectDirectory, PackFile}

The BatchRefUpdate seems to be a little bit quicker, which is good, but
as a separate issue note that https://git.eclipse.org/r/12953 means
the progress is currently incorrectly reported (double-counted).