Address issue #20 - 'LargeObjectException' on huge commit messages
JGit parses Commits (or Tags, or Trees) without streaming, ie in an
in-memory byte-array, and the size of this array is capped by JGit at
the maximum of the WindowCache streaming-threshold, or 5MB, whichever
is larger.

Handling large commits thus requires increasing the WC
streaming-threshold. Previously the WC streaming-threshold was set to
1MB by The BFG, as a performance-improving tweak - we don't necessarily
want to bulk-load the large blobs that are also swept up by this
setting, it slows down content-filtering.

The new flag is '--repo-contains-massive-non-file-objects', ie:

$ bfg --strip-blobs-bigger-than 1K --repo-contains-massive-non-file-objects 20M