# This is a combination of 2 commits. # The first commit's message is:
Fix #23 - stack overflow on ref-update for chromium/src.git

The end-of-run ref-update (performed after all the cleaned objects have
been generated) was **crashing** with a stackoverflow on this repo:

http://git.chromium.org/chromium/src.git

The cause was this code in the JGit library, which tries to work out if
the ref-change is a non-fast-forward update or not:

https://github.com/eclipse/jgit/blob/9b26e4/org.eclipse.jgit/src/org/eclipse/jgit/transport/ReceiveCommand.java#L292-L299

...determining the isMergedInto() boolean is vitally important if you're
fetching updates from someone else's repository, but it's no value to us
at all when we're doing the BFG update (where we /know/ we're going to
'throw away' large amounts of history) - and the calculation can be very
heavy on repos with large numbers of commits (ie ~50s on a 2013 MacBook
Air for a minimal test-case repo with 100000 commits).

Happily, we can avoid this unnecessary calculation by *telling*
JGit that the update is a non-fast-forward (slightly fiddly,
'cos the JGit API didn't anticipate that usage, but possible). The
ref-update is now ~20ms on the same hardware, and the code-path that
crashed the BFG on chromium/src.git is no longer called.

This is lovely! (the BFG just got faster, again), and thanks to
@eglaysher for reporting the issue - it was a great little bug.

https://github.com/rtyley/bfg-repo-cleaner/issues/23

# This is the 2nd commit message:

salad and chips