Fix #23 - stack overflow on ref-update for chromium/src.git
The end-of-run ref-update (performed after all the cleaned objects have
been generated) was **crashing** with a stackoverflow on this repo:

http://git.chromium.org/chromium/src.git

The cause was this code in the JGit library, which tries to work out if
the ref-change is a non-fast-forward update or not:

https://github.com/eclipse/jgit/blob/9b26e4/org.eclipse.jgit/src/org/eclipse/jgit/transport/ReceiveCommand.java#L292-L299

...determining the isMergedInto() boolean is vitally important if you're
fetching updates from someone else's repository, but it's no value to us
at all when we're doing the BFG update (where we /know/ we're going to
'throw away' large amounts of history) - and the calculation can be very
heavy on repos with large numbers of commits (ie ~150s on a 2013 MacBook
Air for a minimal test-case repo with 300000 empty commits).

Unfortunately, there's currently no way to stop JGit from calling
revWalk.isMergedInto() (for instance, by *telling* JGit that the update
is a non-fast-forward) but we /can/ shortcut the work involved by
overriding the method and using our knowledge that 'cleaned' object-ids
are non-ff updates. The ref-update is now ~20ms on the same hardware,
and the code-path that crashed the BFG on chromium/src.git is no longer
called.

Thanks to @eglaysher for reporting the issue - nice bug!

https://github.com/rtyley/bfg-repo-cleaner/issues/23