Add option to fix tree-blob filename collision corruption
The CocoaPods Specs repo (https://github.com/CocoaPods/Specs) suffered
corruption due to an issue in libgit2, resulting in an identically named
tree and blob occupying the same folder:

https://github.com/libgit2/libgit2/pull/2085#issuecomment-33632988
http://blog.cocoapods.org/Repairing-Our-Broken-Specs-Repository/

Hopefully this issue has not affected many more repositories, but it was
interesting to see how history in affected repositories might be
re-written to remove the corruption.

Prior to this change, The BFG was not reliable at removing this kind of
corruption (using ie '--delete-files 2.0.0'). Due to the way it models Git
trees (a model that doesn't permit duplicate filenames) it couldn't see
that the affected tree contained a file with the supplied name (as it was
masked by the identically-named folder) and so didn't generate a cleaned
tree.