treeblobs don't need to be cleaned with explicit concurrency...
The shallow-stack stuff doesn't apply to them, as Git repo folder depth is
rarely more than 100 (opposed to commit histories, which might be as deep
as 1000000 commits or more).

CPU cores will never be starved of work, as there are always other commits
and commit-trees to clean - no need to share cleaning of a signle tree.