Fix #37 - NullPointerException on protecting annotated-tags
Many thanks to @sgandon for reporting the issue. It only affects annotated
tags (tags with specific messages attached to them), which are a little
rarer that 'lightweight' tags.

http://git-scm.com/book/ch2-6.html#Annotated-Tags

The NPE was due to carelessly assuming that revTag.getObject() returned
a **parsed** RevObject... actually it doesn't, so null fields on the
returned object abound.