Probable fix for issue #75 - OOM while trying to find biggest blobs
Problem was not calculating the size - windowCursor.getObjectSize() is
actually pretty good at doing that for big delta'd objects without
allocating too much crazy memory - but establishing the object type (for
low size threshold sizes, you might see objects which are not blobs- there
are some massive commit messages out there for instance). This was done
by fully expanding the delta'd object, and checking the type on the
objectLoader - can't find a nicer way to do this with the existing JGit
lib, but I think in principle it's unnecessary. In any case, for the short
term, lets just assume that any object over 1MB is probably a blob, and
not open it.

This code is all a bit scruffy, but to be fair, I am drunk.

https://github.com/rtyley/bfg-repo-cleaner/issues/75