Publish the BFG's CLI usage text as a Maven artifact for #117
This code just runs the BFG with no arguments, grabs the output to a file,
and adds that to the list of file artifacts that are published when we do
a release. This is prompted by Sebastian Schuberth's request in #117, where
he pointed out that installing the BFG, just to find out what features it
supports, is a cumbersome process for some new users.

Note that the usage text does not show _all_ options - the *really* obscure
ones are marked 'hidden' to reduce the amount of noise users have to see.
The ultimate reference, for all options including 'hidden' options, would
be to look at the `CLIConfig` class itself:

https://github.com/rtyley/bfg-repo-cleaner/blob/master/bfg/src/main/scala/com/madgag/git/bfg/cli/CLIConfig.scala

https://github.com/rtyley/bfg-repo-cleaner/issues/117