Switch to Scalatest from specs2
The ultimate motivation for this switch was the way specs2
handled the new "should not have a StackOverflowError
cleaning a repo with deep history" test (see
`ObjectIdCleanerSpec` in this commit). The test is *meant*
to fail (at least for the moment, because there's no solution
in place at `ObjectIdCleaner` level to address the stack
problem) - but when it fails it produces an exception with
a very deep stack exception, which specs2 chokes on - and
scalatest does not. I did take a bit of a look at fixing
specs2, but in the end, due to various other niggles I've
had with specs2, I decided to move away from it instead,
in favour of Scalatest.

The "deep history" test in `ObjectIdCleanerSpec` is currently
'ignored', precisely because it does fail - the issue is currently
addressed by carefully walking commits in the safest order to avoid
this problem, although it does sometimes go wrong, see eg:

https://github.com/rtyley/bfg-repo-cleaner/issues/174
https://github.com/rtyley/bfg-repo-cleaner/issues/115