Encourage people to not give up in the face of Trump
This doesn't relate to the functionality of the BFG, and some might find
the message unwanted. Yet Trump and his administration are a massive
threat to the future of humanity, and they have the power to repeat their
lies, hate and bile - loudly and incessantly. Occasional reminders that
they're full of shit are a good thing.