Trump is out
This reverts PR #198 ("Encourage people to not give up in the face of
Trump"). Trump is no longer the President of the United States - that
doesn't mean everything is right with the world. But it's at least the
end of this particular global nightmare; 4 years where fascism, misogyny,
lies and hatred were enthusiastically spread by the elected head of the
US government. Good riddance.